fix(db): adjust schema import order
