+++ src/db/index.ts
@@ -2,8 +2,8 @@ import "dotenv/config";
 
 import { drizzle } from "drizzle-orm/node-postgres";
 
-import schema from "./schema";
+import { schema, relations } from "./schema";
 
 export const db = drizzle(process.env.DATABASE_URL!, {
-  schema,
+  schema: { ...schema, ...relations },
 });
