+++ src/db/schema/index.ts
@@ -1,0 +1,4 @@
+export * from "./relations";
+export { default as relations } from "./relations";
+export * from "./tables";
+export { default as schema } from "./tables";
